Moves RustFuncs into a dedicated module
