+++ src/vm/jit/mod.rs
@@ -1,23 +1,22 @@
+use self::funcs::RustFuncs;
 use super::OP_VARARGPREP;
 use crate::lobject::Proto;
 use crate::lstate::CallInfo;
 use crate::ltm::luaT_adjustvarargs;
 use crate::{Lua, LuaFn, StackValue, Thread};
 use alloc::boxed::Box;
-use alloc::vec::Vec;
 use core::mem::{offset_of, transmute};
 use core::ops::Deref;
 use core::pin::Pin;
 use core::ptr::null_mut;
 use core::task::{Context, Poll};
 use cranelift_codegen::ir::types::I32;
-use cranelift_codegen::ir::{
-    AbiParam, ExtFuncData, ExternalName, FuncRef, Function, InstBuilder, MemFlags, Signature, Type,
-    UserExternalName,
-};
+use cranelift_codegen::ir::{AbiParam, FuncRef, Function, InstBuilder, MemFlags, Signature, Type};
 use cranelift_codegen::isa::CallConv;
 use cranelift_frontend::{FunctionBuilder, Variable};
 use target_lexicon::Triple;
+
+mod funcs;
 
 pub async unsafe fn run<A>(
     td: &Thread<A>,
@@ -295,44 +294,6 @@ impl<'a> Compiler<'a> {
         let v = self.fb.ins().iadd(v, f);
 
         self.fb.def_var(self.base, v);
-    }
-}
-
-/// Contains Rust functions that can be called from jitted function.
-#[derive(Default)]
-struct RustFuncs(Vec<*const u8>);
-
-impl RustFuncs {
-    unsafe fn import(
-        &mut self,
-        fb: &mut FunctionBuilder,
-        params: &[Type],
-        output: Option<Type>,
-        f: *const u8,
-    ) -> FuncRef {
-        // Build signature.
-        let mut sig = Signature::new(CallConv::triple_default(&HOST));
-
-        for p in params {
-            sig.params.push(AbiParam::new(*p));
-        }
-
-        if let Some(v) = output {
-            sig.returns.push(AbiParam::new(v));
-        }
-
-        // Import function.
-        let sig = fb.func.import_signature(sig);
-        let name = UserExternalName::new(0, self.0.len().try_into().unwrap());
-        let name = fb.func.declare_imported_user_function(name);
-
-        self.0.push(f);
-
-        fb.func.import_function(ExtFuncData {
-            name: ExternalName::User(name),
-            signature: sig,
-            colocated: false,
-        })
     }
 }
